Ensure cosistent split of description for webhook (#2)
* Ensure cosistent split of description for webhook

Previously, it wasn't possible to reliably split the 
description into name and message, since the 
name could include `:`s.  
This change prepends the length of the `name` 
followed by a pipe.  
Splitting along the first pipe in the string now yields:
1. The length of the name (let's call it `len`)
2. The name, immediately followed by the message 
    (let's call it `joint_description`)
Now the webhook receiver can set `name` to be
`joint_description[:len]`, and `message` to be 
`joint_descritption[len:]`

* Update README.md